moved vs code to its own module and did some other refactoring
